Add documentation for shema type helpers
